refactor: move statistics skeleton components
